Long is fixed and adding InitialData and UserController

+++ src/main/java/org/duze/duzekino/repository/UserRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
-    Long countById(Integer id);
+    Long countById(Long id);
 
     User findUserByUsername(String username);
 }
